requirements.txt
fasttext 
selenium 
bs4 
indic-nlp-library